Spring - Fix bug card service

+++ src/main/java/com/cs/ku/present/service/CardService.java
@@ -1,5 +1,6 @@
 package com.cs.ku.present.service;
 
+import com.cs.ku.present.constant.Status;
 import com.cs.ku.present.dto.mapper.CardMapper;
 import com.cs.ku.present.dto.request.CardRequest;
 import com.cs.ku.present.entity.Card;
@@ -36,7 +37,6 @@ public class CardService {
 
     public Card save(CardRequest request) {
         Card card = CardMapper.toCard(request);
-        card.setId(UUID.randomUUID());
         card.setCreatedTimestamp(ZonedDateTime.now(systemClock));
         card.setCreatedBy(SYSTEM);
         card.setUpdatedTimestamp(ZonedDateTime.now(systemClock));
@@ -46,18 +46,31 @@ public class CardService {
     }
 
     public Card update(String id, CardRequest request) {
-        Card card = CardMapper.toCard(request);
-        card.setId(UUID.fromString(id));
-        card.setCreatedTimestamp(ZonedDateTime.now(systemClock));
-        card.setCreatedBy(SYSTEM);
+        Card card = cardRepository.findById(UUID.fromString(id))
+                .orElseThrow(() -> new NotFoundException(NOT_FOUND.formatted("Card", id)));
+
+        card.setTitle(request.title());
+        card.setDescription(request.description());
+        card.setStatus(Status.fromString(request.status()));
         card.setUpdatedTimestamp(ZonedDateTime.now(systemClock));
         card.setUpdatedBy(SYSTEM);
 
         return cardRepository.save(card);
     }
 
-    public void deleteById(String id) {
-        cardRepository.deleteById(UUID.fromString(id));
+    public void deleteById(String id, boolean isHardDelete) {
+        if (isHardDelete) {
+            cardRepository.deleteById(UUID.fromString(id));
+        } else {
+            Card card = cardRepository.findById(UUID.fromString(id))
+                    .orElseThrow(() -> new NotFoundException(NOT_FOUND.formatted("Card", id)));
+
+            card.setStatus(Status.INACTIVE);
+            card.setUpdatedTimestamp(ZonedDateTime.now(systemClock));
+            card.setUpdatedBy(SYSTEM);
+
+            cardRepository.save(card);
+        }
     }
 
 }
